moved parity tests from ff to ec/hashing/tests.rs and updated cargo files accordingly with test_curves dev dependencies; some errors due to cofactor/cofactor_inv
