+++ ec/Cargo.toml
@@ -26,6 +26,7 @@ zeroize = { version = "1", default-features = false, features = ["zeroize_derive
 [dev-dependencies]
 hashbrown = { version = "0.11.1" }
 blake2 = { version = "0.9.2", default-features = false }
+ark-test-curves = { version = "^0.3.0", path = "../test-curves", default-features = false, features = ["bls12_381_curve"] }
 
 [features]
 default = []
