+++ ec/src/hashing/tests.rs
@@ -1,7 +1,7 @@
 use crate::{
     hashing::{
         curve_maps::{
-            swu::{SWUMap, SWUParams},
+            swu::{SWUMap, SWUParams, parity},
             wb::{WBMap, WBParams},
         },
         field_hashers::DefaultFieldHasher,
@@ -21,6 +21,10 @@ use ark_ff::{
 use ark_ff::SquareRootField;
 use ark_std::vec::Vec;
 use hashbrown::HashMap;
+
+use ark_test_curves::{
+    bls12_381::{Fq, Fq2, Fq6},
+};
 
 pub type F127 = Fp64<F127Parameters>;
 
@@ -420,4 +424,36 @@ fn hash_arbitary_string_to_curve_wb() {
         hash_result.is_on_curve(),
         "hash results into a point off the curve"
     );
+}
+
+#[test]
+fn test_parity() {
+    let a1 = Fq2::new(Fq::from(0), Fq::from(0));
+    let a2 = Fq2::new(Fq::from(0), Fq::from(1));
+    let a3 = Fq2::new(Fq::from(1), Fq::from(0));
+    let a4 = Fq2::new(Fq::from(1), Fq::from(1));
+    let element_test1 = Fq6::new(a1, a2, a3);
+    let element_test2 = Fq6::new(a2, a3, a4);
+    let element_test3 = Fq6::new(a3, a4, a1);
+    let element_test4 = Fq6::new(a4, a1, a2);
+    assert_eq!(parity(&element_test1), false);
+    assert_eq!(parity(&element_test2), false);
+    assert_eq!(parity(&element_test3), true);
+    assert_eq!(parity(&element_test4), true);
+
+    let element_test1 = Fq2::new(Fq::from(0), Fq::from(1));
+    let element_test2 = Fq2::new(Fq::from(1), Fq::from(0));
+    let element_test3 = Fq2::new(Fq::from(10), Fq::from(5));
+    let element_test4 = Fq2::new(Fq::from(5), Fq::from(10));
+    assert_eq!(parity(&element_test1), false);
+    assert_eq!(parity(&element_test2), true);
+    assert_eq!(parity(&element_test3), false);
+    assert_eq!(parity(&element_test4), true);
+
+    let a1 = Fq::from(0);
+    let a2 = Fq::from(1);
+    let a3 = Fq::from(10);
+    assert_eq!(parity(&a1), false);
+    assert_eq!(parity(&a2), true);
+    assert_eq!(parity(&a3), false); 
 }
